feat: return_to path utility refactoring
